requirements.txt
fastapi
uvicorn[standard]
python-dotenv
selenium
selenium-stealth
webdriver-manager
cachetools